+++ cypress/integration/2.3-verify_pickup_points.spec.js
@@ -1,6 +1,6 @@
 import { testSetup, updateRetry } from '../support/common/support'
-import { verifyAllPickUpPoint } from '../support/storelocator.common'
-import { listallPickupPointsAPI } from '../support/store-locator.apis.js'
+import { verifyAllPickUpPoint } from '../support/common.js'
+import { listallPickupPointsAPI } from '../support/api_testcase.js'
 
 describe('Verify all pickup points is displayed', () => {
   // Load test setup
@@ -8,6 +8,7 @@ describe('Verify all pickup points is displayed', () => {
 
   listallPickupPointsAPI()
   it('Verify all pickup points is shown in stores', updateRetry(2), () => {
+    cy.addDelayBetweenRetries(10000)
     verifyAllPickUpPoint()
   })
 })