+++ cypress/integration/2.5-inactive_pickup_scenarios.spec.js
@@ -3,11 +3,12 @@ import {
   updatePickupPointdata,
   listallPickupPointsAPI,
   INTIAL_PICKUP_POINTS_ENV,
-} from '../support/store-locator.apis'
-import { restAPITestCase } from '../support/store-locator.outputvalidation.js'
-import storelocatorSelectors from '../support/storelocator.selectors'
-import { graphql, getStores } from '../support/shipping-policy.graphql'
+} from '../support/api_testcase'
+import { restAPITestCase } from '../support/outputvalidation.js'
+import storelocatorSelectors from '../support/selectors.js'
+import { graphql, getStores } from '../support/graphql_testcase'
 import { storeLocator } from '../support/app_list'
+import { clickLoadAllStores } from '../support/common.js'
 
 const { pickupPoint3Payload } = restAPITestCase
 
@@ -26,9 +27,7 @@ describe('Inactive Pickup Points should not be visible in storefront', () => {
         cy.get(storelocatorSelectors.StorePickUpPointList, {
           timeout: 8000,
         }).should('be.visible')
-        cy.get(storelocatorSelectors.LoadStores, { timeout: 8000 })
-          .should('be.visible')
-          .click()
+        clickLoadAllStores()
         cy.get(storelocatorSelectors.MoreItems, { timeout: 8000 }).should(
           'have.length',
           pickupCount[INTIAL_PICKUP_POINTS_ENV] + 3
@@ -40,14 +39,21 @@ describe('Inactive Pickup Points should not be visible in storefront', () => {
     }
   )
 
-  it('verify getStores with latitude and longitude', updateRetry(3), () => {
+  it('verify getStores with latitude and longitude', updateRetry(5), () => {
     graphql(storeLocator, getStores(-22.94, -43.18), (response) => {
+      cy.addDelayBetweenRetries(4000)
       expect(response.status).to.equal(200)
-      expect(response.body.data.getStores.items.length).to.equal(2)
+      const pickupPoints = response.body.data.getStores.items.map((p) =>
+        p.name.includes('pickup example')
+      )
+
+      expect(pickupPoints.length).to.equal(2)
     })
   })
-  it('verify getStores without latitude and longitude', updateRetry(3), () => {
+
+  it('verify getStores without latitude and longitude', updateRetry(5), () => {
     graphql(storeLocator, getStores(), (response) => {
+      cy.addDelayBetweenRetries(4000)
       expect(response.status).to.equal(200)
       expect(response.body.data.getStores.items.length).to.not.equal(0)
     })