+++ cypress/support/common.js
@@ -1,8 +1,6 @@
 import { updateRetry } from './common/support'
-import storeLocatorSelectors, {
-  findPickupPoint,
-} from './storelocator.selectors'
-import { INTIAL_PICKUP_POINTS_ENV } from './store-locator.apis'
+import storeLocatorSelectors, { findPickupPoint } from './selectors.js'
+import { INTIAL_PICKUP_POINTS_ENV } from './api_testcase'
 
 export function addPickUpPoint(pickPointName) {
   cy.visit('/admin/app/pickup-points')
@@ -29,10 +27,20 @@ export function addPickUpPoint(pickPointName) {
     .contains('Changes saved')
 }
 
+export function clickLoadAllStores() {
+  cy.get('body').then(($body) => {
+    if ($body.find(storeLocatorSelectors.LoadStores).length) {
+      cy.get(storeLocatorSelectors.LoadStores, { timeout: 15000 })
+        .should('be.visible')
+        .click()
+    }
+  })
+}
+
 export function verifyAllPickUpPoint() {
   cy.visitStore()
   cy.get(storeLocatorSelectors.ListOfStores).should('be.visible')
-  cy.get(storeLocatorSelectors.LoadStores).click()
+  clickLoadAllStores()
 
   cy.getPickupPointItem().then((pickupCount) => {
     cy.get(storeLocatorSelectors.MoreItems).should(
@@ -43,18 +51,20 @@ export function verifyAllPickUpPoint() {
       .its('length')
       .then((itemLen) => {
         for (let i = 0; i < itemLen; i++) {
-          cy.get(storeLocatorSelectors.LoadStores).should('be.visible').click()
+          clickLoadAllStores()
           cy.get(storeLocatorSelectors.MoreItems)
             .eq(i)
+            .scrollIntoView()
             .should('be.visible')
             .click()
           cy.get(storeLocatorSelectors.AddressContainer, {
             timeout: 20000,
           }).should('be.visible')
-          cy.get(storeLocatorSelectors.Hours).should('be.visible')
+          cy.get(storeLocatorSelectors.HoursContainer).should('be.visible')
           cy.get(storeLocatorSelectors.BackToPickUpPoint)
             .should('be.visible')
             .click()
+          cy.get(storeLocatorSelectors.MoreItems).should('be.visible')
         }
       })
   })