+++ CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Increased hard retries to verify pickup points in cypress tests
+
+## [0.10.14] - 2022-10-07
+
+### Changed
+
+- GitHub reusable workflow and Cy-Runner update to version 2
+- Updated descriptions for block props in README
+- Moved error logging for `getStores` to GraphQL resolver instead of client
+
+## [0.10.13] - 2022-08-09
+
+## [0.10.12] - 2022-08-04
+
+### Added
+
+- Add holiday exceptions
+
+### Fixed
+
+- Destructuring error in vbase `alreadyHasSitemap` check
+- Limit logistics pickupPoint paging to 99 pages
+- Add `max-height` CSS property to stores list
+
 ## [0.10.11] - 2022-04-28
 
 ### Added